+++ CHANGELOG.rst
@@ -6,28 +6,79 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_
 and this project adheres to `Semantic Versioning <http://semver.org/>`_.
 
-`Unreleased v3.1.0 <https://github.com/cppmicroservices/cppmicroservices/tree/development>`_ (2017-XX-XX)
+`Unreleased v3.2.0 <https://github.com/cppmicroservices/cppmicroservices/tree/development>`_ (2017-XX-XX)
 ---------------------------------------------------------------------------------------------------------
 
-`Full Changelog <https://github.com/cppmicroservices/cppmicroservices/compare/v3.0.0...development>`_
+`Full Changelog <https://github.com/cppmicroservices/cppmicroservices/compare/v3.1.0...development>`_
 
 Added
-~~~~~
+-----
+
+Changed
+-------
+
+Removed
+-------
+
+Deprecated
+----------
+
+Fixed
+-----
+
+`v3.1.0 <https://github.com/cppmicroservices/cppmicroservices/tree/v3.1.0>`_ (2017-06-01)
+-----------------------------------------------------------------------------------------
+
+`Full Changelog <https://github.com/cppmicroservices/cppmicroservices/compare/v3.0.0...v3.1.0>`_
 
 Changed
 ~~~~~~~
 
+- Improved BadAnyCastException message. `#181 <https://github.com/CppMicroServices/CppMicroServices/issues/181>`_
+- Support installing bundles that do not have .DLL/.so/.dylib file extensions. `#205 <https://github.com/CppMicroServices/CppMicroServices/issues/205>`_
+
 Deprecated
 ~~~~~~~~~~
 
-Removed
-~~~~~~~
+- The following BundleContext member functions:
+
+  * ``RemoveBundleListener``
+  * ``RemoveFrameworkListener``
+  * ``RemoveServiceListener``
+
+  And the variants of
+
+  * ``AddBundleListener``
+  * ``AddFrameworkListener``,
+  * ``AddServiceListener``
+
+  that take member functions.
+
+- The free functions:
+
+  * ``ServiceListenerMemberFunctor``
+  * ``BundleListenerMemberFunctor``
+  * ``BindFrameworkListenerToFunctor``
+
+- The functions
+
+  * ``ShrinkableVector::operator[std::size_t]``
+  * ``ShrinkableMap::operator[const Key&]``
+
 
 Fixed
 ~~~~~
 
-Security
-~~~~~~~~
+-  Cannot add more than one listener if its expressed as a lambda.
+   `#95 <https://github.com/CppMicroServices/CppMicroServices/issues/95>`_
+-  Removing Listeners does not work well
+   `#83 <https://github.com/CppMicroServices/CppMicroServices/issues/83>`_
+-  Crash when trying to acquire bundle context
+   `#172 <https://github.com/CppMicroServices/CppMicroServices/issues/172>`_
+-  Fix for ``unsafe_any_cast``
+   `#198 <https://github.com/CppMicroServices/CppMicroServices/pull/198>`_
+-  Stopping a framework while bundle threads are still running may deadlock
+   `#210 <https://github.com/CppMicroServices/CppMicroServices/issues/210>`_
 
 `v3.0.0 <https://github.com/cppmicroservices/cppmicroservices/tree/v3.0.0>`_ (2017-02-08)
 -----------------------------------------------------------------------------------------